resovle code quality issues
