+++ docs/references/configs/page.md
@@ -455,6 +455,9 @@ This document lists all the configuration options supported by the GoFr framewor
 -  MQTT_KEEP_ALIVE
 -  Sends regular messages to check the link is active. May not work as expected if handling func is blocking execution
 
+- MQTT_RETRIEVE_RETAINED
+- Retrieve retained messages on subscription
+
 {% /table %}
 
 **NATS JetStream**
